+++ packages/app/src/lib/Github/index.ts
@@ -1,3 +1,2 @@
-export * from './types';
 export * from './github';
 export * from './github.module';