+++ packages/app/src/lib/Github/types/index.ts
@@ -1,2 +1,0 @@
-export * from './Release';
-export * from './ReleaseAsset';